Removed Handover from ATC activity
