+++ app/Console/Commands/UpdateAtcHours.php
@@ -4,7 +4,6 @@ namespace App\Console\Commands;
 
 use anlutro\LaravelSettings\Facade as Setting;
 use App;
-use App\Models\Handover;
 use App\Models\User;
 use App\Models\AtcActivity;
 use App\Helpers\Vatsim;
@@ -57,8 +56,8 @@ class UpdateAtcHours extends Command
 
         // Fetch members
         $optionalUserIdFilter = $this->argument('user');
-        $handoverMembers = Handover::getRatedMembers($optionalUserIdFilter);
-        $members = User::whereIn('id', $handoverMembers->pluck('id'))->get();
+        $ratedMembers = User::getRatedMembers($optionalUserIdFilter);
+        $members = User::whereIn('id', $ratedMembers->pluck('id'))->get();
 
         // Update member hours
         $this->updateMemberATCHours($members);
